``` 添加外部配置寄存器和相应的回调机制
在Axis类中实现外部配置寄存器，包括一组新的配置项和回调函数，以支持通过CAN接口进行的外部设备配置。此外，更新can_encos.cpp以处理新的CAN消息协议，该协议允许设置和确认外部配置寄存器的值。

在.vscode/settings.json中添加新的C++标准库头文件路径，以支持在 IntelliSense 中对新增的C++标准库头文件的识别。
```

+++ Firmware/MotorControl/encoder.hpp
@@ -10,7 +10,7 @@ class Encoder : public ODriveIntf::EncoderIntf {
 public:
     static constexpr uint32_t MODE_FLAG_ABS = 0x100;
     static constexpr uint32_t MODE_FLAG_485_ABS = 0x200;
-    static constexpr float HALF_CPR =  (1<<18) / 2.0f;
+    static constexpr int32_t HALF_CPR =  (1<<18) / 2;
 
     struct Config_t {
         Mode mode = MODE_INCREMENTAL;
